refetch in updateAppt.js uncommented

+++ client/src/components/updateAppointment/CancelAppt.js
@@ -2,31 +2,40 @@ import React from "react";
 import Navbar from "../Navbar";
 import Footer from "../Footer";
 import { useState } from "react";
-import { useMutation } from "@apollo/client";
-import { useNavigate } from "react-router-dom";
+import { useQuery, useMutation } from "@apollo/client";
 // calling the mutation fom the mutatuion folder
 import { DELETE_APPT } from "../../utils/mutations";
+import { useNavigate } from "react-router-dom";
+import { VIEW_APPOINTMENT } from "../../utils/queries";
 
 const CancelAppt = () => {
   const navigate = useNavigate();
+
   const [confirmationInput, setConfirmationInput] = useState({
     referencenumber: "",
   });
-  const [deleteAppointment, { data, loading, error }] =
-    useMutation(DELETE_APPT);
+  const [deleteAppointment] = useMutation(DELETE_APPT);
+
+  const { data, loading, error, refetch } = useQuery(VIEW_APPOINTMENT, {
+    variables: { appointmentId: confirmationInput.referencenumber },
+  });
 
   const handleInputChange = (e) => {
     const { name, value } = e.target;
-
     console.log("Did the state update?", confirmationInput);
-
     setConfirmationInput({ ...confirmationInput, [name]: value });
+    refetch({ appointmentId: value });
     console.log(confirmationInput);
   };
 
   const deleteFormSubmit = async (e) => {
     e.preventDefault();
     console.log("deleting appointment with id:", confirmationInput.name);
+
+    if (!data || !data.appointment) {
+      alert("there is no appointment with this confirmation number!");
+      return;
+    }
     try {
       const response = await deleteAppointment({
         variables: { appointmentId: confirmationInput.referencenumber },
@@ -44,9 +53,9 @@ const CancelAppt = () => {
   return (
     <div>
       <Navbar />
-      <section className="page-title bg-1 bgSerg">
+      <div className="page-title bg-1 bgSerg">
         <div className="overlay" />
-      </section>
+      </div>
       <section className="appoinment section ">
         <div className="container">
           <div className="row">
@@ -69,11 +78,14 @@ const CancelAppt = () => {
 
                 <form onSubmit={deleteFormSubmit}>
                   <div className="mb-3">
-                    <label className="form-label">
+                    <label htmlFor="exampleInputEmail1" className="form-label">
                       Enter your confirmation number:{" "}
                     </label>
                     <input
                       type="text"
+                      className="form-control"
+                      id="exampleInputEmail1"
+                      aria-describedby="emailHelp"
                       name="referencenumber"
                       value={confirmationInput.referencenumber}
                       onChange={handleInputChange}
